Move get_vm_settings() to HostPluginProtocol
